Change instructions and CI to `npm ci` and commit api/package-lock
With `npm install`, npm installs the most up-to-date packages that are
compatible with the requirements laid out in `package.json`.
The `package-lock.json` file is a complete record of the exact package
versions installed on the device, and it is modified by `npm insall`.
`npm ci` instead installs the exact package versions listed in
`package-lock.json`. This is also why `frontend/api/package-lock.json`
should be committed.
All tests pass with the packages recorded in `frontend/api/package-lock.json`.