[front] feat: polls must be configured to allow public comparisons (#727)
Following our "private by default" policy, polls must be explicitly configured
to allow users to set their comparisons public.

A poll with the `comparisonsCanBePublic` option set to true, will have its
entity card displays the public / private checkbox. The front end page 
/ratings/` will also displays the filter by visibility option and update visibility
buttons.

For now it's a front end only configuration, but it will be easy to make it a
back end configuration (a field in the `Poll` model for instance).