Fix youtube recommendations not always displayed
The content script is not always injected when there's a page change on
youtube. So if it was injected on a /watch URL and then not injected
when returning to /, then the recommendations are empty and therefore
not displayed.

The fix loads the recommendations on the root URL if they haven't been
loaded previously.