[front] style: set background to light grey instead of white
The majority of our components used a white background, and the
main background of the application was white too, which made
hard to distinguish foreground components from background ones.

It also removed the benefits of using MUI `<Paper>`, which have
a drop shadow that makes look like a paper displayed on top on
something else. Casted on a white background this shadow is less
clear-cut, and acts more as a simple delimiter rather than a
proper shadow.