[front] style: set background to light grey instead of white (#879)
The majority of our components used a white background, and the
main background of the application was white too. This absence of
contrast made harder to distinguish foreground components from
background ones.

It also removed the benefits of using MUI `<Paper>` and similar
components, which have a drop shadow that makes them look like a
paper displayed on top on something else. Cast on a white background
this shadow is less clear-cut, and acts more as a simple delimiter rather
than a proper invitation to consider the component more important than
underlying surface.

Having different contrast between foreground and background
component helps to understand what's going on on the screen (modals
are usually display on top of a greyed background), and should help
the user navigate in the application.