I rewrote the code to make it more human-readable. I verified that it successfully pass the tests.
There may be a need to go further to avoid ambiguities,
as the word "trust" is used with too many slightly different meanings.

In particular, I would suggest:
1. The user's field "trust_score" could be replaced with "voting_right".
2. The vouch's field "trust_value" could be replaced by "vouch_value".
3. The vouch's field "by" could be replaced by "from".

As a result, while "trust" remains not fully well defined
(currently, in the code, it only matters if it is > 0),
the word "relative_trust" implies a sum of all users' relative trusts equals 1.
Conversely, a user's "voting_right" an take any positive number.
Moreover pretrusted users (by email domains) are guaranteed to have a voting right of at least 1.