[front] fix: error message more accurate in rate-later form
Adding a video to the rate-later list with a malformed video id/uid
triggered the message "This field cannot be empty", which is confusing
because the form field is not empty, it contains value (an malformed
value, but a value still).

By adding `|| formVideo` instead of `|| ''` we delegate to the back end
the responsability of managing the error and returning an appropriate
message.