[back] feat: full-text search index (#1039)
* full-text search (#902)

* GresilleSiffle review comments

* lint errors

* Use a language column, add default config, index optimization...

* lint

* new migration number

* wip: make it possible to reverse the mig 0046

* move random_page_cost config from migration to db connection options

* rename build_search_vector into update_search_vector

to make explicit that this operation writes into the database

* revert single save in refresh_metadata and add comment

* fix weights_sum definition for Mehestan in 'sort_results'

* replace CRLF newlines to LF in test_text_search for consistency

* optimize 'filter_with_text_query' when explicit language filter is defined

* use normalized total_score (after Mehestan scaling) to compute search_score

* comment details

* workaround for bug in django-debug-toolbar with tuple param in raw SQL query

* add a comment about usage of ANY in raw SQL query

* fix metadata values being indexed in SearchVector as raw json instead of text

Django's `SearchVector` casts the expression as text. As an undesirable consequence
for json fields, the value used to build the vector is a raw json representation
of the metadata content, where special characters and newlines are not correctly processed
(e.g "foo\nbar" will be transformed into 2 tokens "foo" and "nbar").
Using `KeyTextTransform` enables to extract to json value as proper text. Technically it
forces the ORM to use the SQL operator `->>` instead of `->`.

See https://www.postgresql.org/docs/13/functions-json.html for more details about these operators.

* fix db user privileges to create extension 'unaccent' from django migrations

Co-authored-by: Gresille&Siffle <39056254+GresilleSiffle@users.noreply.github.com>
Co-authored-by: Adrien Matissart <a@matissart.net>