[front] i18n: make the `fr` translations less verbose
Also add missing dots at the end of few translations.

I used the indicative present tense for the `fr` translations
instead of the past, to make them less verbose. It shouldn't
cause any senmatic problem as if you have vouched for someone
in the past, it means you are currently vouching for them right
now.