[front] fix: make the trust score dialog box agnostic of the poll
The previous description mentioned the poll videos, but the vouching
page is displayed on all polls. To avoid any confusion, I removed
the `one comparison of public video` by `one public comparison`.

The concept of public comparison is already explained elsewhere in the
application.

It's true that currently the export API only return users having a
public comparison in the poll videos, but in the future we could
have more than one poll accepting public comparisons.