[front] fix: avoid using `<Trans>` when non of its feature are used
The translation key `whereGoPublicData` used to require `<Trans>`
to diplay correctly the `<br>` tag included in its translation, but I
think it's not a good reason to use `<Trans>`.

Most of the time, we should avoid structuring the web page in the
translation to avoid bad surprises, and having to remember that a
part of the component visual is in its JSX and another in its translations.

If we need to create a new paragraph, I'd suggest to add an additional
`<Typography>` component instead of tweaking the translations with `<br>`.

Currently the privacy policy is still a bit raw, but we will be able to
improve it properly in the near future.