Fix broken back button when videos are auto filled
Without this change pressing the browser's back button on the comparison
page after the IDs have been auto filled would clear one input and
trigger a new auto fill.

It also changes the behavior when the user clicks on the "Auto" button
or types a new ID but I believe these events should not add a history
entry anyway.