[ext] Remove the header modification
Since 1accca6eeb75a3357d02f95217acd517b16e4a1d the backend adds a
`Content-Security-Policy` header that overrides the `X-Frame-Options`
header and allows frames from youtube. So this header change would only
be needed for browser that do not support the `frame-ancestors` in
`Content-Security-Policy`, so Chrome < 40 (2015) and Firefox < 58
(2018). I've tested on both Firefox and Chrome and the frame is indeed
allowed if the header change is removed.